+++ src/web-components/tooltip/tooltip.ts
@@ -83,6 +83,6 @@ declare global {
   interface HTMLElementTagNameMap {
     "open-tooltip": Tooltip;
     "tool-tip": Tooltip;
-    "g-open-tooltip": Tooltip;
+    "g-tooltip": Tooltip;
   }
 }